requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
redis
python-dotenv
httpx
pytest